requirements.txt
affine==2.4.0
appdirs==1.4.4
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.7
cycler==0.11.0
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Fiona==1.8.22
fonttools==4.38.0
GDAL==3.5.0
gemgis==1.0.3
geopandas==0.12.2
geoserver-rest==2.4.1
idna==3.4
imageio==2.24.0
importlib-metadata==6.0.0
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.6.3
munch==2.5.0
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pooch==1.6.0
pycodestyle==2.10.0
pygdal==3.5.0.11
Pygments==2.14.0
pyparsing==3.0.9
pyproj==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
pyvista==0.37.0
rasterio==1.3.4
requests==2.28.2
scooby==0.7.0
seaborn==0.12.2
shapely==2.0.0
six==1.16.0
snuggs==1.4.7
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
vtk==9.2.5
xmltodict==0.13.0
zipp==3.11.0
